requirements.txt
google-adk==0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
resend==2.9.0
